requirements/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --no-emit-package setuptools --no-strip-extras requirements/requirements.in -o requirements/requirements.txt
amqp==5.3.1
    # via kombu
asgiref==3.8.1
    # via
    #   django
    #   django-allauth
    #   django-htmx
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
billiard==4.2.1
    # via celery
boto3==1.35.60
    # via -r requirements/requirements.in
botocore==1.35.60
    # via
    #   boto3
    #   s3transfer
brotli==1.1.0
    # via whitenoise
celery[redis]==5.4.0
    # via
    #   -r requirements/requirements.in
    #   django-celery-beat
celery-progress==0.4
    # via -r requirements/requirements.in
certifi==2024.8.30
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cron-descriptor==1.4.5
    # via django-celery-beat
cryptography==43.0.3
    # via
    #   fido2
    #   pyjwt
django==5.1.3
    # via
    #   -r requirements/requirements.in
    #   django-allauth
    #   django-anymail
    #   django-celery-beat
    #   django-htmx
    #   django-otp
    #   django-storages
    #   django-template-partials
    #   django-timezone-field
    #   django-waffle
    #   djangorestframework
    #   drf-spectacular
django-allauth[mfa, socialaccount]==65.2.0
    # via -r requirements/requirements.in
django-anymail[mailgun]==12.0
    # via -r requirements/requirements.in
django-celery-beat==2.7.0
    # via -r requirements/requirements.in
django-environ==0.11.2
    # via -r requirements/requirements.in
django-htmx==1.21.0
    # via -r requirements/requirements.in
django-otp==1.5.4
    # via -r requirements/requirements.in
django-storages==1.14.4
    # via -r requirements/requirements.in
django-template-partials==24.4
    # via -r requirements/requirements.in
django-timezone-field==7.0
    # via django-celery-beat
django-waffle==4.1.0
    # via -r requirements/requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements/requirements.in
    #   drf-spectacular
djangorestframework-api-key==3.0.0
    # via -r requirements/requirements.in
drf-spectacular==0.27.2
    # via -r requirements/requirements.in
fido2==1.1.3
    # via django-allauth
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2024.10.1
    # via jsonschema
kombu==5.4.2
    # via celery
oauthlib==3.2.2
    # via requests-oauthlib
packaging==24.2
    # via djangorestframework-api-key
prompt-toolkit==3.0.48
    # via click-repl
psycopg2-binary==2.9.10
    # via -r requirements/requirements.in
pycparser==2.22
    # via cffi
pyjwt[crypto]==2.9.0
    # via django-allauth
python-crontab==3.2.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   celery
    #   python-crontab
pyyaml==6.0.2
    # via drf-spectacular
qrcode==8.0
    # via django-allauth
redis==5.2.0
    # via celery
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   django-allauth
    #   django-anymail
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via django-allauth
rpds-py==0.21.0
    # via
    #   jsonschema
    #   referencing
s3transfer==0.10.3
    # via boto3
six==1.16.0
    # via python-dateutil
sqlparse==0.5.2
    # via django
tzdata==2024.2
    # via
    #   celery
    #   django-celery-beat
    #   kombu
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.2.3
    # via
    #   botocore
    #   django-anymail
    #   requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
whitenoise[brotli]==6.8.2
    # via -r requirements/requirements.in
